Move reusable dialog objects to ama-testing/e2e
