+++ e2e/tests/project/delete-project.e2e.ts
@@ -18,7 +18,7 @@
 import { testConfig } from '../../test.config';
 import { LoginPage, LoginPageImplementation } from 'ama-testing/e2e';
 import { DashboardPage } from '../../pages/dashboard.page';
-import { DeleteEntityDialog } from '../../pages/dialog/delete-entity.dialog';
+import { DeleteEntityDialog } from 'ama-testing/e2e';
 import { SnackBar } from '../../pages/snackbar';
 import { NodeEntry } from 'alfresco-js-api-node';
 import { Backend } from 'ama-testing/e2e';
